add attachments and linked resources to content model
